update todo, budgets, and local data
