+++ Frontend/src/components/TripPageComponents/AccommodationComponent/AccommodationComponent.css
@@ -1,0 +1,169 @@
+/* Accommodation Component Styles */
+.section-card {
+    background-color: #ECE4D7;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    padding: 20px;
+    flex: 1;
+    min-width: 300px;
+  }
+  
+  .section-card h2 {
+    color: #333;
+    margin-top: 0;
+    margin-bottom: 10px;
+  }
+  
+  .section-card h3 {
+    color: #333;
+    margin-top: 15px;
+    margin-bottom: 10px;
+  }
+  
+  .divider {
+    height: 1px;
+    background-color: #ddd;
+    margin-bottom: 15px;
+  }
+  
+  /* Accommodation Table */
+  .accommodation-table {
+    width: 100%;
+    margin-bottom: 15px;
+  }
+  
+  .acc-header {
+    display: flex;
+    background-color: #9FAC9D;
+    color: white;
+    padding: 8px 0;
+    border-radius: 4px 4px 0 0;
+  }
+  
+  .acc-row {
+    display: flex;
+    border-bottom: 1px solid #ddd;
+    padding: 8px 0;
+  }
+  
+  .acc-row:last-child {
+    border-bottom: none;
+  }
+  
+  .acc-cell {
+    flex: 1;
+    padding: 5px 10px;
+    display: flex;
+    align-items: center;
+  }
+  
+  .acc-cell.actions {
+    flex: 0.5;
+    justify-content: center;
+  }
+  
+  .delete-acc-btn {
+    background-color: #CDACA1;
+    color: white;
+    border: none;
+    border-radius: 50%;
+    width: 20px;
+    height: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    font-size: 14px;
+    line-height: 1;
+  }
+  
+  .delete-acc-btn:hover {
+    background-color: #b3978d;
+  }
+  
+  /* Add Form */
+  .add-form {
+    margin-top: 20px;
+    padding-top: 15px;
+    border-top: 1px solid #ddd;
+  }
+  
+  .add-form h3 {
+    margin-top: 0;
+  }
+  
+  .form-row {
+    margin-bottom: 15px;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .form-row label {
+    margin-bottom: 5px;
+    font-weight: bold;
+    color: #444;
+  }
+  
+  .form-row input, .form-row select {
+    padding: 10px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    font-size: 16px;
+  }
+  
+  .styled-select {
+    appearance: none;
+    background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="12" height="6"><path d="M0 0l6 6 6-6z" fill="%23333"/></svg>');
+    background-repeat: no-repeat;
+    background-position: right 10px center;
+    background-size: 10px;
+    padding-right: 30px;
+  }
+  
+  .btn-add {
+    background-color: #9FAC9D;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    padding: 8px 15px;
+    cursor: pointer;
+    font-size: 14px;
+    transition: background-color 0.2s;
+  }
+  
+  .btn-add:hover {
+    background-color: #89999A;
+  }
+  
+  /* Budget Summary */
+  .budget-summary-section {
+    margin-top: 20px;
+    border-top: 1px solid #ddd;
+    padding-top: 15px;
+  }
+  
+  .budget-summary-section h3 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    color: #333;
+  }
+  
+  .budget-item {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 5px;
+  }
+  
+  .budget-item.total {
+    font-weight: bold;
+    margin-top: 10px;
+    padding-top: 10px;
+    border-top: 1px dashed #ddd;
+  }
+  
+  .empty-message {
+    padding: 15px;
+    color: #888;
+    text-align: center;
+    font-style: italic;
+  }
