+++ Frontend/src/components/TripPageComponents/TodoComponent/TodoComponent.css
@@ -1,0 +1,104 @@
+/* Todo Component Styles */
+.section-card {
+    background-color: #ECE4D7;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    padding: 20px;
+    flex: 1;
+    min-width: 300px;
+  }
+  
+  .section-card h2 {
+    color: #333;
+    margin-top: 0;
+    margin-bottom: 10px;
+  }
+  
+  .divider {
+    height: 1px;
+    background-color: #ddd;
+    margin-bottom: 15px;
+  }
+  
+  /* Todo Section */
+  .todo-section {
+    margin-bottom: 20px;
+  }
+  
+  .todo-section h3 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    color: #333;
+  }
+  
+  .todo-list {
+    background-color: white;
+    border-radius: 4px;
+    overflow: hidden;
+    border: 1px solid #ddd;
+    margin-bottom: 10px;
+    min-height: 50px;
+  }
+  
+  .todo-item {
+    display: flex;
+    justify-content: space-between;
+    padding: 10px;
+    border-bottom: 1px solid #eee;
+  }
+  
+  .todo-item:last-child {
+    border-bottom: none;
+  }
+  
+  .add-todo-form {
+    display: flex;
+    gap: 8px;
+    margin-top: 8px;
+  }
+  
+  .add-todo-form input {
+    flex-grow: 1;
+    padding: 8px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+  }
+  
+  .add-todo-form button {
+    background-color: #9FAC9D;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    padding: 8px 15px;
+    cursor: pointer;
+  }
+  
+  .add-todo-form button:hover {
+    background-color: #89999A;
+  }
+  
+  .delete-todo-btn {
+    background-color: #CDACA1;
+    color: white;
+    border: none;
+    border-radius: 50%;
+    width: 20px;
+    height: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    font-size: 14px;
+    line-height: 1;
+  }
+  
+  .delete-todo-btn:hover {
+    background-color: #b3978d;
+  }
+  
+  .empty-message {
+    padding: 15px;
+    color: #888;
+    text-align: center;
+    font-style: italic;
+  }
