+++ Frontend/src/components/TripPageComponents/TripListComponent/TripListComponent.css
@@ -1,42 +1,122 @@
-/* Trip List Component Styles */
-.trip-list {
-    background-color: white;
+/* Trip Table Styles */
+.trip-table {
+    background-color: #ECE4D7;
     border-radius: 8px;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    margin: 20px 0;
-}
-
-.table__header {
-    background-color: #f5f5f5;
-    border-top-left-radius: 8px;
-    border-top-right-radius: 8px;
-}
-
-.t-header__row {
+    margin-bottom: 20px;
+    overflow: hidden;
+  }
+  
+  .table-header {
     display: flex;
-    padding: 12px;
-}
-
-.t-header__col {
+    background-color: #9FAC9D;
+    padding: 15px;
+    font-weight: bold;
+    color: white;
+  }
+  
+  .header-cell {
     flex: 1;
+    padding: 5px 10px;
+  }
+  
+  .header-cell.actions {
+    flex: 0.5;
+    text-align: center;
+  }
+  
+  .table-body {
+    max-height: 300px;
+    overflow-y: auto;
+  }
+  
+  .table-row {
+    display: flex;
+    border-bottom: 1px solid #ddd;
+    padding: 12px 15px;
+    transition: background-color 0.2s;
+  }
+  
+  .table-row:hover {
+    background-color: #f0f0f0;
+    cursor: pointer;
+  }
+  
+  .table-row:last-child {
+    border-bottom: none;
+  }
+  
+  .table-cell {
+    flex: 1;
+    padding: 5px 10px;
+    display: flex;
+    align-items: center;
+  }
+  
+  .table-cell.actions {
+    flex: 0.5;
+    display: flex;
+    justify-content: center;
+    gap: 5px;
+  }
+  
+  .table-cell.trip-name,
+  .table-cell.trip-travelers,
+  .table-cell.trip-dates {
+    cursor: pointer;
+  }
+  
+  .table-cell.trip-name {
     font-weight: bold;
     color: #333;
-}
-
-.table__body {
-    min-height: 100px;
-}
-
-.no-trips {
-    text-align: center;
-    padding: 20px;
+  }
+  
+  .clickable {
+    cursor: pointer;
+  }
+  
+  .clickable:hover {
+    text-decoration: underline;
+  }
+  
+  .edit-trip-btn, .delete-trip-btn {
+    border: none;
+    border-radius: 4px;
+    padding: 5px 10px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+  }
+  
+  .edit-trip-btn {
+    background-color: #9FAC9D;
+    color: white;
+  }
+  
+  .edit-trip-btn:hover {
+    background-color: #89999A;
+  }
+  
+  .delete-trip-btn {
+    background-color: #CDACA1;
+    color: white;
+    border-radius: 50%;
+    width: 25px;
+    height: 25px;
+    padding: 0;
+    font-size: 18px;
+    line-height: 1;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .delete-trip-btn:hover {
+    background-color: #b3978d;
+  }
+  
+  .empty-message {
+    padding: 15px;
     color: #888;
-}
-
-.t-body__row {
-    transition: background-color 0.3s ease;
-}
-
-.t-body__row:hover {
-    background-color: #f0f0f0;
-}
+    text-align: center;
+    font-style: italic;
+  }
