+++ Frontend/src/components/TripPageComponents/TripPageComponent/TripPageComponent.css
@@ -40,431 +40,19 @@
   background-color: #b3978d;
 }
 
-/* Trip Table Styles */
-.trip-table {
-  background-color: #ECE4D7;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin-bottom: 20px;
-  overflow: hidden;
-}
-
-.table-header {
-  display: flex;
-  background-color: #9FAC9D;
-  padding: 15px;
-  font-weight: bold;
-  color: white;
-}
-
-.header-cell {
-  flex: 1;
-  padding: 5px 10px;
-}
-
-.header-cell.actions {
-  flex: 0.5;
-  text-align: center;
-}
-
-.table-body {
-  max-height: 300px;
-  overflow-y: auto;
-}
-
-.table-row {
-  display: flex;
-  border-bottom: 1px solid #ddd;
-  padding: 12px 15px;
-  transition: background-color 0.2s;
-}
-
-.table-row:hover {
-  background-color: #f0f0f0;
-  cursor: pointer;
-}
-
-.table-row:last-child {
-  border-bottom: none;
-}
-
-.table-cell {
-  flex: 1;
-  padding: 5px 10px;
-  display: flex;
-  align-items: center;
-}
-
-.table-cell.actions {
-  flex: 0.5;
-  display: flex;
-  justify-content: center;
-  gap: 5px;
-}
-
-.table-cell.trip-name,
-.table-cell.trip-travelers,
-.table-cell.trip-dates {
-  cursor: pointer;
-}
-
-.table-cell.trip-name {
-  font-weight: bold;
-  color: #333;
-}
-
-.clickable {
-  cursor: pointer;
-}
-
-.clickable:hover {
-  text-decoration: underline;
-}
-
-.edit-trip-btn, .delete-trip-btn {
-  border: none;
-  border-radius: 4px;
-  padding: 5px 10px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.edit-trip-btn {
-  background-color: #9FAC9D;
-  color: white;
-}
-
-.edit-trip-btn:hover {
-  background-color: #89999A;
-}
-
-.delete-trip-btn {
-  background-color: #CDACA1;
-  color: white;
-  border-radius: 50%;
-  width: 25px;
-  height: 25px;
-  padding: 0;
-  font-size: 18px;
-  line-height: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.delete-trip-btn:hover {
-  background-color: #b3978d;
-}
-
-/* Trip Form Styles */
-.trip-form {
-  background-color: #ECE4D7;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  padding: 20px;
-  margin-bottom: 20px;
-}
-
-.trip-form h2 {
-  color: #333;
-  margin-top: 0;
-  margin-bottom: 20px;
-  border-bottom: 1px solid #ddd;
-  padding-bottom: 10px;
-}
-
-.form-row {
-  margin-bottom: 15px;
-  display: flex;
-  flex-direction: column;
-}
-
-.form-row label {
-  margin-bottom: 5px;
-  font-weight: bold;
-  color: #444;
-}
-
-.form-row input, .form-row select {
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 16px;
-}
-
-.form-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 10px;
-  margin-top: 20px;
-}
-
-.btn-cancel, .btn-create, .btn-save, .btn-add {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.2s;
-}
-
-.btn-cancel {
-  background-color: #ddd;
-  color: #333;
-}
-
-.btn-create, .btn-save, .btn-add {
-  background-color: #9FAC9D;
-  color: white;
-}
-
-.btn-add {
-  padding: 8px 15px;
-  font-size: 14px;
-}
-
-.btn-cancel:hover {
-  background-color: #ccc;
-}
-
-.btn-create:hover, .btn-save:hover, .btn-add:hover {
-  background-color: #89999A;
-}
-
-/* Trip Details Section */
-.trip-details-section {
-  margin-top: 20px;
-}
-
-.trip-details-section h2 {
-  margin-bottom: 15px;
-  color: #333;
-  border-bottom: 1px solid #ddd;
-  padding-bottom: 10px;
-}
-
-.flexible-container {
-  display: flex;
-  gap: 20px;
-  flex-wrap: wrap;
-}
-
-.section-card {
-  background-color: #ECE4D7;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  padding: 20px;
-  flex: 1;
-  min-width: 300px;
-}
-
-.section-card h2 {
-  color: #333;
-  margin-top: 0;
-  margin-bottom: 10px;
-}
-
-.section-card h3 {
-  color: #333;
-  margin-top: 15px;
-  margin-bottom: 10px;
-}
-
-.divider {
-  height: 1px;
-  background-color: #ddd;
-  margin-bottom: 15px;
-}
-
-/* Accommodation Table */
-.accommodation-table {
-  width: 100%;
-  margin-bottom: 15px;
-}
-
-.acc-header {
-  display: flex;
-  background-color: #9FAC9D;
-  color: white;
-  padding: 8px 0;
-  border-radius: 4px 4px 0 0;
-}
-
-.acc-row {
-  display: flex;
-  border-bottom: 1px solid #ddd;
-  padding: 8px 0;
-}
-
-.acc-row:last-child {
-  border-bottom: none;
-}
-
-.acc-cell {
-  flex: 1;
-  padding: 5px 10px;
-  display: flex;
-  align-items: center;
-}
-
-.acc-cell.actions {
-  flex: 0.5;
-  justify-content: center;
-}
-
-.delete-acc-btn, .delete-todo-btn {
-  background-color: #CDACA1;
-  color: white;
-  border: none;
-  border-radius: 50%;
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  font-size: 14px;
-  line-height: 1;
-}
-
-.delete-acc-btn:hover, .delete-todo-btn:hover {
-  background-color: #b3978d;
-}
-
-/* Add Form */
-.add-form {
-  margin-top: 20px;
-  padding-top: 15px;
-  border-top: 1px solid #ddd;
-}
-
-.add-form h3 {
-  margin-top: 0;
-}
-
-/* Budget Summary */
-.budget-summary-section {
-  margin-top: 20px;
-  border-top: 1px solid #ddd;
-  padding-top: 15px;
-}
-
-.budget-summary-section h3 {
-  margin-top: 0;
-  margin-bottom: 10px;
-  color: #333;
-}
-
-.budget-item {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 5px;
-}
-
-.budget-item.total {
-  font-weight: bold;
-  margin-top: 10px;
-  padding-top: 10px;
-  border-top: 1px dashed #ddd;
-}
-
-/* Todo Section */
-.todo-section {
-  margin-bottom: 20px;
-}
-
-.todo-section h3 {
-  margin-top: 0;
-  margin-bottom: 10px;
-  color: #333;
-}
-
-.todo-list {
-  background-color: white;
-  border-radius: 4px;
-  overflow: hidden;
-  border: 1px solid #ddd;
-  margin-bottom: 10px;
-  min-height: 50px;
-}
-
-.todo-item {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  border-bottom: 1px solid #eee;
-}
-
-.todo-item:last-child {
-  border-bottom: none;
-}
-
-.add-todo-form {
-  display: flex;
-  gap: 8px;
-  margin-top: 8px;
-}
-
-.add-todo-form input {
-  flex-grow: 1;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-.add-todo-form button {
-  background-color: #9FAC9D;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  padding: 8px 15px;
-  cursor: pointer;
-}
-
-.add-todo-form button:hover {
-  background-color: #89999A;
-}
-
-.empty-message {
-  padding: 15px;
-  color: #888;
-  text-align: center;
-  font-style: italic;
-}
-
-/* Save Button Container */
-.save-button-container {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 20px;
-}
-
-.btn-save {
-  background-color: #89999A;
-  color: white;
-  font-size: 16px;
-  padding: 12px 25px;
-}
-
-.btn-save:hover {
-  background-color: #778889;
-}
-
 /* Responsive Design */
 @media (max-width: 768px) {
-  .flexible-container {
+  .trip-page {
+    padding: 10px;
+  }
+  
+  .trip-page__header {
     flex-direction: column;
+    gap: 10px;
+    padding: 10px;
   }
   
-  .section-card {
-    min-width: unset;
-  }
-  
-  .add-todo-form {
-    flex-direction: column;
-  }
-  
-  .add-todo-form > * {
-    margin-bottom: 8px;
+  .trip-page__header h1 {
+    font-size: 20px;
   }
 }
